-fixed few input issues. - added dynamic loading of textboxes based on cell selected in the datagrid.

+++ Coursework2/invoice.xaml.cs
@@ -23,12 +23,11 @@ namespace Coursework2
         {
             InitializeComponent();
         }
-        public invoice(List<int> totalprices) :this()
+        public invoice(List<double> totalprices) :this()
         {
             //totalprices = { guestsPerNight, extraBF, extraM, extraCH, noOfNights };
             try
             {
-
                 textBox_guestPerNight.Text = "£ " + totalprices[0];
                 textBox_guestTotal.Text = "£ " + totalprices[0] * totalprices[4];
                 textBox_breakfast.Text = "£ " + totalprices[1];
